broadcast: Added full support for schedules batch
+ Updating a schedule belonging to a batch also updates the other schedules in it
+ New `remove_from_batch` property when updating a schedule to remove it from its batch
+ New route to list campaigns by Ids

+++ Modules/Broadcast/Models/Schedule.php
@@ -143,8 +143,10 @@ class Schedule extends BroadcastResourceModel {
         parent::boot();
 
         static::deleting(static function (Schedule $schedule) {
-            // Clean up all reviews for the schedule
-            $schedule->reviews()->delete();
+            if ($schedule->isForceDeleting()) {
+                // Clean up all reviews for the schedule
+                $schedule->reviews()->delete();
+            }
 
             // Dispatch job to delete schedules in broadcasters
             DeleteScheduleJob::dispatch($schedule->getKey());
